Centralize testing configurations
This centralizes the testing config in testing.gradle, moving
the handling of SKIP_FDB_TESTS and SKIP_SLOW_TESTS there.
And:
- adds a default timeout to all tests of 2 seconds if you're
  skipping slow tests or 30 seconds if not
- adds a performanceTest task with the hopes of removing the
  process detecting code used before
- rearranges the tests for BlockingInAsyncDetection to make sure
  that the lucene configuration doesn't break the default configuration
- disables assertions and BlockingInAsyncDetection for performance
  tests
- removes jenkinsUnitTest; it wasn't used by anyone
