+++ fdb-record-layer-lucene/fdb-record-layer-lucene.gradle
@@ -71,33 +71,16 @@ if (protoMajorVersion == '2') {
     }
 }
 
-def skipFDB = System.getenv('SKIP_FDB_TESTS') != null && System.getenv('SKIP_FDB_TESTS') == 'true'
-def skipSlow = System.getenv('SKIP_SLOW_TESTS') != null && System.getenv('SKIP_SLOW_TESTS') == 'true'
-test {
-    useJUnitPlatform {
-        // Used by lucene tests to create Directory objects for testing our codecs
-        systemProperty 'tests.directory', 'com.apple.foundationdb.record.lucene.codec.TestFDBDirectory'
-        // uncomment this line, and comment out the one above to easily test against Lucene's RAMDirectory to see how
-        // it might differ
-        //systemProperty 'tests.directory', 'org.apache.lucene.store.RAMDirectory'
-        // Uncomment below line to run nightly tests
-        // systemProperty 'tests.nightly', 'true'
-        if (skipFDB) {
-            excludeTags 'RequiresFDB'
-        }
-        if (skipSlow) {
-            excludeTags 'Slow'
-        }
+// Disable blocking detection, because we do this *a lot* in lucene
+test.systemProperties['com.apple.foundationdb.record.blockingInAsyncDetection'] = System.getenv('BLOCKING_DETECTION') ?: "DISABLED"
+// Used by lucene tests to create Directory objects for testing our codecs
+test.systemProperties['tests.directory'] = 'com.apple.foundationdb.record.lucene.codec.TestFDBDirectory'
+// uncomment this line, and comment out the one above to easily test against Lucene's RAMDirectory to see how
+// it might differ
+// test.systemProperties['tests.directory'] = 'org.apache.lucene.store.RAMDirectory'
 
-        // We need the destructive tests to be available in the 'test' task, so that IntelliJ's gradle test runner can
-        // find them. However, we _only_ want to run them as part of the 'test' task if we're running from IntelliJ.
-        // This isn't a great way to detect it, but it's the only one I could find. If this ever breaks, I'd look here first.
-        if (System.getenv('XPC_SERVICE_NAME') == null || !(System.getenv('XPC_SERVICE_NAME').contains("intellij") || System.getenv('XPC_SERVICE_NAME').contains("idea"))) {
-            excludeTags 'WipesFDB'
-            excludeTags 'Performance'
-        }
-    }
-}
+// Uncomment below line to run nightly tests
+// test.systemProperties['tests.nightly'] = 'true'
 
 publishing {
     publications {
